Move "def popget" from utils.py to app/models/models_utils.py
